Updated human ocr

+++ kaylee/client/kaylee.coffee
@@ -17,27 +17,33 @@
 #     worker : null # Worker object
 #     subscribed : false
 
+AUTO_MODE = 0x2
+MANUAL_MODE = 0x4
+
 kl.config = {}
+# The MANUAL_MODE project's namespace.
+# (Note: AUTO_MODE project namespace is defined in klworker.coffee).
+kl.pj = {}
 
 kl.api =
     register : () ->
         kl.get("/kaylee/register",
                kl.node_registered.trigger,
                kl.server_raised_error.trigger)
-        return null
+        return
 
     subscribe : (name) ->
         kl.post("/kaylee/apps/#{name}/subscribe/#{kl.node_id}",
                 null,
                 kl.node_subscribed.trigger,
                 kl.server_raised_error.trigger)
-        return null
+        return
 
     get_action : () ->
         kl.get("/kaylee/actions/#{kl.node_id}",
                kl.action_received.trigger,
                kl.server_raised_error.trigger)
-        return null
+        return
 
     send_result : (results) ->
         kl.post("/kaylee/actions/#{kl.node_id}", results,
@@ -47,7 +53,7 @@ kl.api =
             ),
             kl._default_server_error_handler
         )
-        return null
+        return
 
 kl.register = () ->
     if kl._test_node().features.worker
@@ -55,7 +61,7 @@ kl.register = () ->
     else
         kl.log("Node cannot be registered: client does not meet the "
                "requirements.")
-    return null
+    return
 
 kl.subscribe = (name) ->
     kl.app = {
@@ -65,21 +71,21 @@ kl.subscribe = (name) ->
         subscribed : false
     }
     kl.api.subscribe(name)
-    return null
+    return
 
 kl.get_action = () ->
     if kl.app.subscribed
         kl.api.get_action()
-    return null
+    return
 
 kl.send_result = (data) ->
     if kl.app.subscribed
         kl.api.send_result(data)
-    return null
+    return
 
 kl._message_to_worker = (msg, data = {}) ->
     kl.app.worker.postMessage({'msg' : msg, 'data' : data})
-    return null
+    return
 
 kl._default_server_error_handler = (err) ->
     kl.server_raised_error.trigger(err)
@@ -93,62 +99,73 @@ on_node_registered = (data) ->
     for key, val of data.config
         kl.config[key] = val
     kl.node_id = data.node_id
-    return null
+    return
 
 on_node_subscribed = (config) ->
     kl.app.config = config
-    kl.app.worker.terminate() if kl.app.worker?
-    kl.app.subscribed = true
+    switch config.__kl_project_mode__
+        when AUTO_MODE
+            kl.app.worker.terminate() if kl.app.worker?
 
-    worker = new Worker(kl.config.WORKER_SCRIPT);
-    kl.app.worker = worker;
-    worker.addEventListener('message', ((e) -> on_worker_message(e.data)),
-                            false);
-    worker.addEventListener('error', ((e) -> on_worker_error(e)), false);
-    kl._message_to_worker('import_project', {
-        'kl_config' : kl.config,
-        'app_config' : kl.app.config,
-    })
-    return null
+            worker = new Worker(kl.config.WORKER_SCRIPT)
+            kl.app.worker = worker
+            worker.onmessage = (e) -> worker_message_handler(e)
+            worker.onerror = kl.worker_raised_error.trigger
+            kl._message_to_worker('import_project', {
+                'kl_config' : kl.config,
+                'app_config' : kl.app.config,
+            })
+
+            kl.app.subscribed = true
+
+        when MANUAL_MODE
+            kl.include(config.__kl_project_script__, () ->
+                pj.init(kl.config, kl.app.config)
+            )
+        else
+            kl.error('Unknown kaylee Project mode')
+
+    return
+
+
 
 on_node_unsubscibed = (data) ->
     kl.app.subscribed = false
     kl.app.worker.terminate()
     kl.app.worker = null;
-    return null
+    return
 
 on_project_imported = () ->
     kl.get_action()
-    return null
+    return
 
 on_action_received = (data) ->
     switch data.action
         when 'task' then kl.task_received.trigger(data.data)
         when 'unsubscribe' then kl.node_unsubscibed.trigger(data.data)
         else alert("Unknown action: #{data.action}")
-    return null
+    return
 
 on_task_received = (data) ->
     kl._message_to_worker('solve_task', data)
-    return null
+    return
 
 on_task_completed = (data) ->
     kl.send_result(data)
-    return null
+    return
+
 
 # Kaylee worker event handlers
-on_worker_message = (data) ->
+worker_message_handler = (event) ->
+    data = event.data
     msg = data.msg
     mdata = data.data
     switch msg
         when '__klw_log__' then kl.log.trigger(mdata)
         when 'project_imported' then kl.project_imported.trigger(mdata)
         when 'task_completed' then kl.task_completed.trigger(mdata)
-    return null
+    return
 
-on_worker_error = (e) ->
-    kl.worker_raised_error.trigger(e)
-    return null
 
 # Kaylee events
 kl.node_registered = new Event(on_node_registered)
@@ -160,6 +177,7 @@ kl.task_received = new Event(on_task_received)
 kl.task_completed = new Event(on_task_completed)
 kl.log = new Event()
 kl.result_sent = new Event()
+kl.error = new Event()
 kl.worker_raised_error = new Event(kl.error)
 kl.server_raised_error = new Event(kl.error)
